Move NquadsUdParser to a separate package
