+++ timbuctoo-instancev4/src/main/java/nl/knaw/huygens/timbuctoo/v5/rdfio/implementations/rdf4j/Rdf4jIoFactory.java
@@ -4,7 +4,7 @@ import nl.knaw.huygens.timbuctoo.v5.filestorage.dto.CachedLog;
 import nl.knaw.huygens.timbuctoo.v5.rdfio.RdfIoFactory;
 import nl.knaw.huygens.timbuctoo.v5.rdfio.RdfParser;
 import nl.knaw.huygens.timbuctoo.v5.rdfio.RdfSerializer;
-import nl.knaw.huygens.timbuctoo.v5.rdfio.implementations.rdf4j.NquadsUdParser.NquadsUdParserFactory;
+import nl.knaw.huygens.timbuctoo.v5.rdfio.implementations.rdf4j.parsers.NquadsUdParser.NquadsUdParserFactory;
 import org.eclipse.rdf4j.rio.RDFFormat;
 import org.eclipse.rdf4j.rio.RDFParserRegistry;
